[#20] SidebarTab 타입 파일 이전
